pytests/wireless_radiotap: make Intel-specific config optional
The 'bf_params' file is Intel-specific and should not cause a
non-Intel-WiFi device to fail the test. It appears to be useful for
helping ensure (more-)reliable monitor-mode capture of beacons, and so
removing its usage entirely would negatively impact the speed with which
an Intel device could complete this test.

Let's just make it optional, based on whether the file exists. This
should allow the test to pass when using non-Intel WiFi options.

Also drop the Intel-specific documentation note, since it's been wrong
for years.

BUG=b:140426523, b:74049380
TEST=`make test`
TEST=run it

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/1781669
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Tested-by: Chia-Hsiu Chang <chia-hsiu.chang@quanta.corp-partner.google.com>
Tested-by: Brian Norris <briannorris@chromium.org>
Auto-Submit: Brian Norris <briannorris@chromium.org>

+++ py/test/pytests/wireless_radiotap.py
@@ -2,7 +2,7 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-"""A factory test for checking Wifi antenna for Intel WP2 7260 chip.
+"""A factory test for checking Wifi antennas.
 
 Description
 -----------
@@ -299,11 +299,14 @@ class Capture(object):
   def set_beacon_filter(self, value):
     """Sets beacon filter.
 
-    This function may only for Intel WP2 7260 chip.
+    This function is currently only needed for Intel WiFi.
     """
     path = '/sys/kernel/debug/ieee80211/%s/netdev:%s/iwlmvm/bf_params' % (
         self.phy, self.parent_device)
-    self.dut.WriteFile(path, 'bf_enable_beacon_filter=%d\n' % value)
+    if self.dut.path.exists(path):
+      session.console.info('Setting beacon filter (enable=%d) for Intel WiFi',
+                           value)
+      self.dut.WriteFile(path, 'bf_enable_beacon_filter=%d\n' % value)
 
   def __enter__(self):
     if not self.created_device:
